moved config file back from .papers/config.json to .papersconfig.json
This is an reversal of some earlier changes, to prepare for moving the
backup directory in some central, safe location, and keep the local
config file. As a consequence, the config file won't be tracked in git
anymore. On the other hand, an encompassing project can
now track .papersconfig.json while ignoring the .papers folder.