Centralized backup repository
Despite what I first argued in #51, I now consider the central backup
option to be the safest. More background work with git is needed to keep
the size of the backup repo reasonable (e.g. remove what's older than
X), but that is not urgent, by definition, and manual maintenance is
possible (like removing the whole thing, since the backup dir is
indicated in the config anyway).

TODO:
- initialize back-up git repo upon saving in case it was removed, or to
  make it work even without install.